refactor: move watch v3 data response interface to api wrapper library
