--- packages/lib/api-wrapper/src/types.ts
@@ -1,6 +1,0 @@
-export type VideoId = `${"so" | "sm" | "nm"}${number}`;
-export type ChannelId = `ch${number}`;
-
-export interface ApiEndpoints {
-  endpoint: URL;
-}
