+++ packages/lib/api-wrapper/src/types/index.ts
@@ -1,0 +1,3 @@
+export * from "./api";
+export * from "./video-info";
+export * from "./watch-data";
